:sparkles: feature: S3 스토리지에 2개 이상의 객체 삭제 기능 구현 및 주석 추가 (#18)
