--- src/main/java/com/simplelog/api/commons/file/S3Uploader.java
@@ -1,79 +1,0 @@
-package com.simplelog.api.commons.file;
-
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.DeleteObjectRequest;
-import com.amazonaws.services.s3.model.PutObjectRequest;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.util.Optional;
-
-@Slf4j
-@RequiredArgsConstructor
-@Component
-public class S3Uploader {
-
-    private final AmazonS3Client amazonS3Client;
-
-    @Value("${cloud.aws.s3.bucket}")
-    public String bucket;
-
-    /***************************************************************************************************
-     * Upload
-     **************************************************************************************************/
-
-    public String upload(MultipartFile multipartFile, String dirName, String fileName) throws IOException {
-        File uploadFile = convert(multipartFile)
-                .orElseThrow(() -> new IllegalArgumentException("error: MultipartFile -> File convert fail"));
-
-        return uploadToS3(uploadFile, dirName, fileName);
-    }
-
-    private String uploadToS3(File uploadFile, String dirName, String fileName) {
-        String savedFileName = dirName + "/" + fileName;
-        System.out.println("savedFileName = " + savedFileName);
-
-        String uploadImageUrl = putS3(uploadFile, savedFileName);
-        removeTempFile(uploadFile);
-        return uploadImageUrl;
-    }
-
-    private String putS3(File uploadFile, String fileName) {
-        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile).withCannedAcl(CannedAccessControlList.PublicRead));
-        return amazonS3Client.getUrl(bucket, fileName).toString();
-    }
-
-    private void removeTempFile(File targetFile) {
-        if (targetFile.delete()) {
-            log.info("파일이 삭제되었습니다.");
-            return;
-        }
-        log.info("파일이 삭제되지 못했습니다.");
-    }
-
-    private Optional<File> convert(MultipartFile file) throws IOException {
-        File convertFile = new File(System.getProperty("user.dir") + "/" + file.getOriginalFilename());
-        if (convertFile.createNewFile()) {
-            try (FileOutputStream fos = new FileOutputStream(convertFile)) {
-                fos.write(file.getBytes());
-            }
-            return Optional.of(convertFile);
-        }
-
-        return Optional.empty();
-    }
-
-    /***************************************************************************************************
-     * Remove
-     **************************************************************************************************/
-    public void remove(String dirName, String fileName) {
-        amazonS3Client.deleteObject(new DeleteObjectRequest(bucket + dirName, fileName));
-    }
-}
